Restructure communication crate
