+++ crates/communication/src/client/mod.rs
@@ -1,0 +1,13 @@
+mod communication;
+mod connector;
+mod id_tracker;
+mod output_subscription_manager;
+mod parameter_subscription_manager;
+mod receiver;
+mod requester;
+mod responder;
+mod types;
+
+pub use crate::client::communication::Communication;
+pub use connector::ConnectionStatus;
+pub use types::{Cycler, CyclerOutput, HierarchyType, Output, OutputHierarchy, SubscriberMessage};
