+++ crates/communication/src/lib.rs
@@ -1,14 +1,3 @@
-mod communication;
-mod connector;
-mod id_tracker;
-mod output_subscription_manager;
-mod parameter_subscription_manager;
-mod receiver;
-mod requester;
-mod responder;
+pub mod client;
+pub mod messages;
 pub mod server;
-mod types;
-
-pub use crate::communication::Communication;
-pub use connector::ConnectionStatus;
-pub use types::{Cycler, CyclerOutput, HierarchyType, Output, OutputHierarchy, SubscriberMessage};
